More unsigned int fixes.
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@1106 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ org/xbill/DNS/DSRecord.java
@@ -85,8 +85,8 @@ throws IOException
 {
 	DSRecord rec = new DSRecord(name, dclass, ttl);
 	rec.footprint = st.getUInt16();
-	rec.alg = (byte) st.getUInt8();
-	rec.digestid = (byte) st.getUInt8();
+	rec.alg = st.getUInt8();
+	rec.digestid = st.getUInt8();
 
 	// note that the draft says that the digest is presented as hex,
 	// not base64.
